requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pycryptodomex==3.20.0
python-decouple==3.8
python-telegram-bot==21.4
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
steam==1.4.4
urllib3==2.2.2
vdf==3.4
